+++ tests/apps/dummy/bearer.py
@@ -2,7 +2,7 @@ from starlette.authentication import SimpleUser
 
 from starlette_auth_toolkit.base.backends import BearerAuthBackend
 
-from ..utils import make_get_app
+from ..utils import get_base_app
 
 TOKEN = "s3kr3t_t0k3n"
 
@@ -15,4 +15,4 @@ class BearerAuth(BearerAuthBackend):
 
 
 def get_app():
-    return make_get_app(backend=BearerAuth())
+    return get_base_app(backend=BearerAuth())